+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Bugfix: fix a leak in the rsusermap code (#467)
 - Bugfix: use single-line WTOs for single-line messages (#509)
 - Bugfix: make sure modreg-based modules are never deleted (#517, zowe/zss#749)
+- Feature: add a modify command for displaying the module registry information
+  (#519)
 
 ## `3.1.0`
 - Enhancement: `configmgr extract` option supports simple array (#502)